amari.xlog: Emit config_get after every sync(attached) instead of only after every attach
We emit config_get after every attach from the beginning of xlog in
e0cc8a38 (amari.xlog: Initial draft). The reasoning here is that it is
useful by default to know configuration of a service.

In the previous patch we added sync events so that xlog stream becomes
self-synchronizing. To continue that line it is now useful to have that
config_get emitted after every such synchronization point instead of
only after attaching to the service. That's what hereby patch does.

As a bonus the code is reworked in a way that config_get setup is not
hardcoded anymore and config_get periodicity now can be controlled by
users via explicitly specifying config_get in the logspec.